Add back html.js
