+++ src/html.js
@@ -1,0 +1,35 @@
+import React from 'react';
+
+const JS_NPM_URLS = [
+  '//unpkg.com/docsearch.js@2.4.1/dist/cdn/docsearch.min.js',
+];
+
+export default class HTML extends React.Component {
+  render() {
+    return (
+      <html lang="en" {...this.props.htmlAttributes}>
+        <head>
+          {JS_NPM_URLS.map(url => (
+            <link key={url} rel="preload" href={url} as="script" />
+          ))}
+          <meta charSet="utf-8" />
+          <meta httpEquiv="X-UA-Compatible" content="IE=edge" />
+          <meta
+            name="viewport"
+            content="width=device-width, initial-scale=1.0"
+          />
+          <link rel="icon" href="/favicon.ico" />
+          {this.props.headComponents}
+        </head>
+        <body {...this.props.bodyAttributes}>
+          <div
+            id="___gatsby"
+            dangerouslySetInnerHTML={{__html: this.props.body}}
+          />
+          {this.props.postBodyComponents}
+          {JS_NPM_URLS.map(url => <script key={url} src={url} />)}
+        </body>
+      </html>
+    );
+  }
+}
